+++ scripts/base_network.py
@@ -4,7 +4,7 @@
 
 # coding: utf-8
 """
-Creates the network topology from a `ENTSO-E map extract <https://github.com/PyPSA/GridKit/tree/master/entsoe>`_ (January 2020) as a PyPSA network.
+Creates the network topology from a `ENTSO-E map extract <https://github.com/PyPSA/GridKit/tree/master/entsoe>`_ (March 2022) as a PyPSA network.
 
 Relevant Settings
 -----------------
@@ -41,7 +41,7 @@ Relevant Settings
 Inputs
 ------
 
-- ``data/entsoegridkit``:  Extract from the geographical vector data of the online `ENTSO-E Interactive Map <https://www.entsoe.eu/data/map/>`_ by the `GridKit <https://github.com/pypsa/gridkit>`_ toolkit dating back to January 2020.
+- ``data/entsoegridkit``:  Extract from the geographical vector data of the online `ENTSO-E Interactive Map <https://www.entsoe.eu/data/map/>`_ by the `GridKit <https://github.com/martacki/gridkit>`_ toolkit dating back to March 2022.
 - ``data/parameter_corrections.yaml``: Corrections for ``data/entsoegridkit``
 - ``data/links_p_nom.csv``: confer :ref:`links`
 - ``data/links_tyndp.csv``: List of projects in the `TYNDP 2018 <https://tyndp.entsoe.eu/tyndp2018/>`_ that are at least *in permitting* with fields for start- and endpoint (names and coordinates), length, capacity, construction status, and project reference ID.
@@ -165,8 +165,9 @@ def _load_links_from_eg(buses, eg_links):
 
     links['length'] /= 1e3
 
-    # hotfix
-    links.loc[links.bus1=='6271', 'bus1'] = '6273'
+    # Skagerrak Link is connected to 132kV bus which is removed in _load_buses_from_eg.
+    # Connect to neighboring 380kV bus
+    links.loc[links.bus1=='6396', 'bus1'] = '6398'
 
     links = _remove_dangling_branches(links, buses)