+++ doc/release_notes.rst
@@ -26,6 +26,8 @@ Upcoming Release
 
 * Network building is made deterministic by supplying a fixed random state to network clustering routines.
 
+* New network topology extracted from the ENTSO-E interactive map.
+
 
 PyPSA-Eur 0.4.0 (22th September 2021)
 =====================================